doge_pi/requirements.txt
opencv-python-headless
Pillow